test: a user can not delete himself

+++ medical-appointment-app/app/Http/Controllers/Admin/UserController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Spatie\Permission\Models\Role;
 
 class UserController extends Controller
@@ -112,6 +113,19 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
+        //No permitir que el usuario logueado se borre a si mismo
+        if ($user->id == Auth::user()->id){
+            session()->flash('swal',[
+                'icon' => 'error',
+                'tittle' => 'Acción denegada',
+                'text' => 'No puedes eliminar a ti mismo',
+            ]);
+            abort(403, 'No puedes borrar a tu propio usuario');
+
+            return redirect(route('admin.users.index'));
+        }
+
+
         //Eliminar roles asociados a un usuario
         $user->roles()->detach();
